design pages

+++ app/Http/Controllers/Backend/CustomerController.php
@@ -23,7 +23,7 @@ class CustomerController extends Controller
     public function index()
     {
         if (auth()->check()) {
-            if (auth()->user()->can('primary-client')) {
+            if (auth()->user()->can('client')) {
                 $customers = Customer::all();
                 $divisions = Division::all();
                 $serviceCategories = ServiceCategory::all();
@@ -40,18 +40,16 @@ class CustomerController extends Controller
 
     public function getdata(Request $request)
     {
-        if ($request->status == null) {
-            $status = 0;
-        } else {
-            $status = $request->input('status'); // Default to 0 if not provided
-        }
+        $status = $request->input('status'); // Default to 0 if not provided
         $division_id = $request->input('division_id');
         $district_id = $request->input('district_id');
         $location_id = $request->input('location_id');
         $service_category_id = $request->input('service_category');
         $project_id = $request->input('project_id');
 
-        $model = Customer::where('status',0)->get();
+        $model = Customer::orderBy('created_at', 'desc');
+
+        // dd($status);
         if ($status) {
             $model->when($status, function ($query) use ($status) {
                 $query->where('status', $status);
@@ -105,9 +103,10 @@ class CustomerController extends Controller
                     $deleteBtn = '';
                     $addContactClient = '';
                     $addNonProspective = '';
+                    $viewConversation = '';
 
-                    $editUrl = route('primary-clients.edit', $row->id);
-                    $deleteUrl = route('primary-clients.distroy', $row->id);
+                    $editUrl = route('all-clients.edit', $row->id);
+                    $deleteUrl = route('all-clients.distroy', $row->id);
                     $csrfToken = csrf_field();
                     $method = method_field('DELETE');
 
@@ -139,22 +138,22 @@ class CustomerController extends Controller
                         </svg>
                         </span></button>';
                     }
-
-                    $viewConversation = '<button class="btn btn-sm ms-2 view-conversation" style="padding: 8px; background-color:#6c757d" data-customer-id="' . $row->id . '"  data-bs-toggle="modal" data-bs-target="#viewConversationdrop">
+                    if (auth()->user()->can('view-conversation')) {
+                        $viewConversation = '<button class="btn btn-sm ms-2 view-conversation" style="padding: 8px; background-color:#6c757d" data-customer-id="' . $row->id . '"  data-bs-toggle="modal" data-bs-target="#viewConversationdrop">
                     <span>
                     <svg fill="#ffffff" height="20px" width="20px" version="1.1" id="Capa_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 490 490" xml:space="preserve" stroke="#ffffff"><g id="SVGRepo_bgCarrier" stroke-width="0"></g><g id="SVGRepo_tracerCarrier" stroke-linecap="round" stroke-linejoin="round"></g><g id="SVGRepo_iconCarrier"> <g> <g> <g> <circle cx="367.5" cy="85.6" r="51.3"></circle> <path d="M367.5,160.6c-67.7,0-122.5,54.8-122.5,122.5h245C490,215.4,435.2,160.6,367.5,160.6z"></path> <circle cx="122.5" cy="292.1" r="51.3"></circle> <path d="M122.5,367.5C54.8,367.5,0,422.3,0,490h245C245,422.3,190.2,367.5,122.5,367.5z"></path> <path d="M217,180.4V131h31.1V0H22.6v131.1h142.7L217,180.4z M196,55.6h5.1c5.4,0,10.1,4.3,10.1,10.1s-4.3,10.1-10.1,10.1H196 c-5.4,0-10.1-4.3-10.1-10.1S190.6,55.6,196,55.6z M74.7,75.4h-5.1c-5.4,0-10.1-4.3-10.1-10.1c0-5.8,4.3-10.1,10.1-10.1h5.1 c5.4,0,10.1,4.3,10.1,10.1C84.8,71.2,80.1,75.4,74.7,75.4z M137.7,75.4h-5.1c-5.4,0-10.1-4.3-10.1-10.1c0-5.8,4.3-10.1,10.1-10.1 h5.1c5.4,0,10.1,4.3,10.1,10.1C147.8,71.2,143.1,75.4,137.7,75.4z"></path> </g> </g> </g> </g></svg>
                     </span></button>';
+                    }
 
 
-
-                    if (auth()->user()->can('primary-client-edit')) {
+                    if (auth()->user()->can('client-edit')) {
                         $editBtn = '<a href="' . $editUrl . '" class="edit btn btn-sm btn-success me-2 rounded" style="padding:8px;"><span>' .
                             '<svg viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" style="width: 20px; height: 20px;">' .
                             '<g id="SVGRepo_iconCarrier"> <path fill-rule="evenodd" clip-rule="evenodd" d="M21.1213 2.70705C19.9497 1.53548 18.0503 1.53547 16.8787 2.70705L15.1989 4.38685L7.29289 12.2928C7.16473 12.421 7.07382 12.5816 7.02986 12.7574L6.02986 16.7574C5.94466 17.0982 6.04451 17.4587 6.29289 17.707C6.54127 17.9554 6.90176 18.0553 7.24254 17.9701L11.2425 16.9701C11.4184 16.9261 11.5789 16.8352 11.7071 16.707L19.5556 8.85857L21.2929 7.12126C22.4645 5.94969 22.4645 4.05019 21.2929 2.87862L21.1213 2.70705ZM18.2929 4.12126C18.6834 3.73074 19.3166 3.73074 19.7071 4.12126L19.8787 4.29283C20.2692 4.68336 20.2692 5.31653 19.8787 5.70705L18.8622 6.72357L17.3068 5.10738L18.2929 4.12126ZM15.8923 6.52185L17.4477 8.13804L10.4888 15.097L8.37437 15.6256L8.90296 13.5112L15.8923 6.52185ZM4 7.99994C4 7.44766 4.44772 6.99994 5 6.99994H10C10.5523 6.99994 11 6.55223 11 5.99994C11 5.44766 10.5523 4.99994 10 4.99994H5C3.34315 4.99994 2 6.34309 2 7.99994V18.9999C2 20.6568 3.34315 21.9999 5 21.9999H16C17.6569 21.9999 19 20.6568 19 18.9999V13.9999C19 13.4477 18.5523 12.9999 18 12.9999C17.4477 12.9999 17 13.4477 17 13.9999V18.9999C17 19.5522 16.5523 19.9999 16 19.9999H5C4.44772 19.9999 4 19.5522 4 18.9999V7.99994Z" fill="#ffffff"></path> </g></svg>' .
                             '</span></a>';
                     }
 
-                    if (auth()->user()->can('primary-client-delete')) {
+                    if (auth()->user()->can('client-delete')) {
                         $deleteBtn = '<form action="' . $deleteUrl . '" method="POST">
                     ' . $csrfToken . '
                     ' . $method . '
@@ -168,11 +167,6 @@ class CustomerController extends Controller
 
                     return '<div class="d-flex align-items-center justify-content-center mb-2">'
                         . $editBtn . $deleteBtn . $addConversation . $viewConversation .
-                        '</div>'
-                        . '<div class="mt-2 d-flex align-items-center justify-content-center">'
-                        . $addContactClient  .
-                        '</div>' . '<div class="mt-2 d-flex align-items-center justify-content-center">'
-                        . $addNonProspective  .
                         '</div>';
                 })
                 ->rawColumns(['action'])
@@ -196,7 +190,7 @@ class CustomerController extends Controller
     public function create()
     {
         if (auth()->check()) {
-            if (auth()->user()->can('primary-client-create')) {
+            if (auth()->user()->can('client-create')) {
                 $divisions = Division::all();
                 $projects = Project::all();
                 $locations = Location::all();
@@ -214,8 +208,6 @@ class CustomerController extends Controller
 
     public function store(Request $request)
     {
-        // dd($request->all());
-        // Validate the request data
         $request->validate([
             'name' => 'nullable|string',
             'company_name' => 'nullable|string',
@@ -241,7 +233,7 @@ class CustomerController extends Controller
         $customer = Customer::create([
             'customer_id' => $newId,
             'name' => $request->input('name'),
-            'status' => 0,
+            'status' => 6,
             'company_name' => $request->input('company_name'),
             'location_id' => $request->input('location_id'),
             'email' => $request->input('email'),
@@ -267,14 +259,14 @@ class CustomerController extends Controller
         }
 
 
-        return redirect()->route('primary-clients')->with('success', 'Data created successfully');
+        return redirect()->route('all-clients')->with('success', 'Data created successfully');
     }
 
 
     public function edit($id)
     {
         if (auth()->check()) {
-            if (auth()->user()->can('primary-client-edit')) {
+            if (auth()->user()->can('client-edit')) {
                 $customer = Customer::with('projects')->findOrFail($id);
                 $locations = Location::all();
                 $divisions = Division::all();
@@ -374,17 +366,26 @@ class CustomerController extends Controller
 
 
         // Redirect back with success message
-        return redirect()->route('primary-clients')->with('success', 'Customer updated successfully');
+        return redirect()->route('all-clients')->with('success', 'Customer updated successfully');
     }
 
 
     public function distroy($id)
     {
-        $customer = Customer::findOrFail($id);
-        CustomerProject::where('customer_id', $customer->id)->delete();
-        ConversationLog::where('customer_id', $customer->id)->delete();
-        $customer->delete();
-        return redirect()->route('primary-clients')->with('success', 'Customer deleted successfully');
+        if (auth()->check()) {
+            if (auth()->user()->can('client-delete')) {
+                $customer = Customer::findOrFail($id);
+                CustomerProject::where('customer_id', $customer->id)->delete();
+                ConversationLog::where('customer_id', $customer->id)->delete();
+                $customer->delete();
+                return redirect()->route('all-clients')->with('success', 'Customer deleted successfully');
+            } else {
+                auth()->logout(); // Log out the user
+                return redirect()->route('login')->with('error', 'You do not have permission to edit and have been logged out.');
+            }
+        } else {
+            return redirect()->back()->with('error', 'You need to login first.');
+        }
     }
 
     public function updateStatus(Request $request, $id)
@@ -446,9 +447,9 @@ class CustomerController extends Controller
         }
     }
 
-    public function getClientDetails($id) {
+    public function getClientDetails($id)
+    {
         $client = Customer::find($id); // Adjust model and logic as per your setup
         return response()->json(['client' => $client]);
     }
-    
 }
